Move PostAttendance into its own file
