+++ src/components/templates/PosterHall/components/PosterAttendance/PosterAttendance.tsx
@@ -1,0 +1,22 @@
+import { faUserFriends } from "@fortawesome/free-solid-svg-icons";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+
+interface PosterAttendanceProps {
+  userCount: number;
+}
+
+export const PosterAttendance: React.FC<PosterAttendanceProps> = ({
+  userCount,
+}) => {
+  if (userCount === 0) return <></>;
+
+  return (
+    <div className="PosterAttendance">
+      <FontAwesomeIcon
+        className="PosterAttendance--icon"
+        icon={faUserFriends}
+      />
+      <span>{userCount}</span>
+    </div>
+  );
+};
