+++ src/components/templates/PosterHall/components/PosterAttendance/index.tsx
@@ -1,0 +1,1 @@
+export { PosterAttendance } from "./PosterAttendance";
