+++ src/components/templates/PosterHall/components/PosterPreview/PosterPreview.scss
@@ -76,20 +76,3 @@ $online-container-min-width: 50px;
     }
   }
 }
-
-.PosterAttendance {
-  margin: 0 0 0 auto;
-  background: var(--content--under-30a);
-  border-radius: $border-radius--xl;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  min-width: $online-container-min-width;
-  height: $spacing--xxl;
-  padding: $spacing--md;
-  font-size: $font-size--sm;
-
-  &--icon {
-    margin-right: $spacing--xs;
-  }
-}
